it-lab-g4-master/work-time-report-backend/src/main/java/com/epolsoft/wtr/controller/BookController.java
package com.epolsoft.wtr.controller;

import com.epolsoft.wtr.entity.Book;
import com.epolsoft.wtr.service.BookService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import java.net.URI;
import java.util.List;

@RestController
@RequestMapping(path = "/book")
public class BookController {

    private final BookService bookService;

    public BookController(BookService bookService) {
        this.bookService = bookService;
    }

    @GetMapping
    public List<Book> getAllBooks() {
        return bookService.findAll();
    }

    @GetMapping("/{id}")
    public Book gtBookById(@PathVariable("id") Integer id) {
        return bookService.findById(id);
    }

    @PostMapping
    public ResponseEntity<Object> addBook(@RequestBody Book book) {
        book = bookService.createOrUpdate(book);
        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                .path("/{id}")
                .buildAndExpand(book.getId())
                .toUri();
        return ResponseEntity.created(location)
                .build();
    }
}
